Finished manual mode: - Refactored networking mechanism - Fixed reserving and freeing robots and controllers - Tied up whole code

+++ robotymobilne_projekt/robotymobilne_projekt/GUI/Converters/ValToPercentConverter.cs
@@ -1,9 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Data;
 
 namespace robotymobilne_projekt.GUI.Converters
@@ -12,9 +8,9 @@ namespace robotymobilne_projekt.GUI.Converters
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if(value is double)
+            if(value is int)
             {
-                double toConvert = (double)value;
+                double toConvert = (int)value;
                 return mapValues(toConvert, 4300, 5000, 0, 100);
             }
             return 0;
